requirements.txt
flask==2.2.2
peewee==3.15.3
psycopg2-binary==2.9.5
requests==2.28.1
flask-restful==0.3.9
pylint==2.15.5
pylint-flask==0.6
beautifulsoup4==4.11.1
pillow==9.2.0
python-dotenv==0.21.0
flask_login==0.6.2
Flask-Limiter==2.7.0
flask-cors==3.0.10
rq==1.11.1
redis==4.3.4
waitress==2.1.2
python-slugify==8.0.1